+++ lib/sbmt/outbox/engine.rb
@@ -26,6 +26,11 @@ module Sbmt
         c.database_switcher = "Sbmt::Outbox::DatabaseSwitcher"
         c.batch_process_middlewares = []
         c.item_process_middlewares = []
+
+        if defined?(::Sentry)
+          c.batch_process_middlewares.push("Sbmt::Outbox::Middleware::Sentry::TracingBatchProcessMiddleware")
+          c.item_process_middlewares.push("Sbmt::Outbox::Middleware::Sentry::TracingItemProcessMiddleware")
+        end
       end
 
       rake_tasks do